requirement.txt
fastapi
uvicorn
mysql-connector-python
pycryptodome